Refactor tests to be format specific
